ml/requirements.txt
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
pillow==10.1.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
aiofiles==23.2.1